fido: Make user verification availability synchronous
This function is intended to signal whether user verification may potentially
be available for the authenticator, not whether user verification will be
available for a particular request. Since this can be known ahead of time for
our clients, there is no need for this function to be async. Making this async
also aligns us closer with upstream.

Per-request availability checks should be done in the caller's implementation
of check_user().