[PM-29535] Update cargo-sort (#602)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-29535

## 📔 Objective

Upgrades the version of cargo-sort used to the latest. This changes
slightly the ordering and formatting of some of the Cargo.toml files,
but they should be functionally identical. The change should help reduce
the ammount of accidental changes made by devs when they install a too
new version of cargo-sort.

If you're running cargo-sort locally, make sure it's updated and that
you're running the correct command (Now it needs a `--grouped`
parameter). I'll post this on the SDK channel as well once the change is
merged:
```sh
# To install it
cargo install cargo-sort --version 2.0.2 --locked

# To sort and format the cargo manifests
cargo sort --workspace --grouped

# To check if the files are sorted correctly, without making any changes
cargo sort --workspace --grouped --check
```

## 🚨 Breaking Changes

<!-- Does this PR introduce any breaking changes? If so, please describe
the impact and migration path for clients.

If you're unsure, the automated TypeScript compatibility check will run
when you open/update this PR and provide feedback.

For breaking changes:
1. Describe what changed in the client interface
2. Explain why the change was necessary
3. Provide migration steps for client developers
4. Link to any paired client PRs if needed

Otherwise, you can remove this section. -->

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes